+++ packages/vexide-devices/src/smart/gps.rs
@@ -122,8 +122,8 @@ impl GpsSensor {
 }
 
 impl SmartDevice for GpsSensor {
-    fn port_index(&self) -> u8 {
-        self.port.index()
+    fn port_number(&self) -> u8 {
+        self.port.number()
     }
 
     fn device_type(&self) -> SmartDeviceType {