+++ packages/vexide-devices/src/smart/optical.rs
@@ -184,8 +184,8 @@ impl OpticalSensor {
 }
 
 impl SmartDevice for OpticalSensor {
-    fn port_index(&self) -> u8 {
-        self.port.index()
+    fn port_number(&self) -> u8 {
+        self.port.number()
     }
 
     fn device_type(&self) -> SmartDeviceType {