+++ CHANGELOG.md
@@ -32,8 +32,12 @@ Before releasing:
 ### Changed
 
 - Overhauled the design of the startup banner.
-- `DistanceSensor::distance` now returns an `Option` that will be `None` if the sensor is out of range.
-- Adjusted distance sensor status code errors to be more clear.
+- `DistanceSensor::distance` now returns an `Option` that will be `None` if the sensor is out of range. (#113) (**Breaking Change**)
+- Adjusted distance sensor error names. (#113) (**Breaking Change**)
+- Renamed `SmartDevice::port_index` and `SmartPort::index` to `SmartDevice::port_number` and `SmartPort::port_number`. (#121) (**Breaking Change**)
+- Renamed `AdiDevice::port_index` and `AdiPort::index` to `AdiDevice::port_number` and `AdiDevice::port_number`. (#121) (**Breaking Change**)
+- `SmartPort::device_type` now no longer returns a `Result`. (#121) (**Breaking Change**)
+- Marks many futures as `#[must_use]` to warn when futures are created without `await`ing them. (#112)
 
 ### Removed